+++ Source/OxyPlot.GtkSharp/PlotView.cs
@@ -262,16 +262,11 @@ namespace OxyPlot.GtkSharp
                 if (this.currentModel != null)
                 {
                     ((IPlotModel)this.currentModel).AttachPlotView(null);
+                    this.currentModel = null;
                 }
 
                 if (this.Model != null)
                 {
-                    if (this.Model.PlotView != null)
-                    {
-                        throw new InvalidOperationException(
-                            "This PlotModel is already in use by some other plot view.");
-                    }
-
                     ((IPlotModel)this.Model).AttachPlotView(this);
                     this.currentModel = this.Model;
                 }