+++ Source/OxyPlot.Xamarin.Mac/PlotView.cs
@@ -98,11 +98,6 @@ namespace OxyPlot.Xamarin.Mac
 
                     if (value != null)
                     {
-                        if (value.PlotView != null)
-                        {
-                            throw new InvalidOperationException("This PlotModel is already in use by some other PlotView control.");
-                        }
-
                         ((IPlotModel)value).AttachPlotView(this);
                         this.model = value;
                     }